Cumulus instance passed to bench
Added new Cumulus class that stores the reference to the Connection.
An instance of Cumulus is passed to the bench function and it's
reference to the connection is used for patching. Users should now
either initialize a global instance of Cumulus via initCumulus() and the
bench function will use the global instance by default or an alternative
is to pass a Cumulus instance manual which may be used if multiple
connections are required.